dkImageLayoutInitialize: Implement rendering to 3D slices
Implement rendering to 3D image slices. The GPU supports rendering to 2D
slices of 3D images as if it were a 2D array. This also includes
rendering to different layers through gl_Layer.

Rendering mipmaps and to different layers through gl_Layer has been
tested. gl_Layer's access is robust, applications are not able to write
out of bounds and the layer is clamped to the highest valid slice.